requirements.txt
absolufy-imports==0.3.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.6.2
asgiref==3.8.1
attrs==25.3.0
authlib==1.3.2
black==24.8.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
contourpy==1.1.1
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.11
dj-database-url==1.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
docstring-parser==0.17.0
dotenv==0.9.9
drf-yasg==1.21.10
fastapi==0.116.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.90.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.30.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==8.5.0
inflection==0.5.1
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langcodes==3.4.1
language-data==1.3.0
mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
mypy-extensions==1.1.0
narwhals==1.42.1
networkx==3.1
numpy==1.24.3
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.2.1
setuptools==75.3.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.3.1
sniffio==1.3.1
spacy==4.0.0.dev2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
srsly==2.4.8
sse-starlette==2.1.3
starlette==0.44.0
sympy==1.13.3
tenacity==8.5.0
thinc==9.0.0.dev5
threadpoolctl==3.5.0
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typer==0.17.4
typing-extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
websockets==13.1
whitenoise==6.7.0
wrapt==1.17.3
zipp==3.20.2